requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
Markdown==3.3.4
PyJWT==2.0.1
pytz==2021.1
requests==2.25.1
idna
sqlparse==0.4.1
urllib3==1.26.4
gunicorn==20.1.0
python-dotenv==0.16.0
psycopg2==2.8.6
